move all the things
